+++ runUnitTest.sh
@@ -2,7 +2,7 @@
 
 # MIT License
 #
-# (C) Copyright [2021-2022] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2021-2022,2024] Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -33,7 +33,7 @@ echo "COMPOSE_FILE: $COMPOSE_FILE"
 
 
 function cleanup() {
-  docker-compose down
+  docker compose down
   if ! [[ $? -eq 0 ]]; then
     echo "Failed to decompose environment!"
     exit 1
@@ -43,8 +43,8 @@ function cleanup() {
 
 
 echo "Starting containers..."
-docker-compose build --no-cache
-docker-compose up --exit-code-from unit-tests unit-tests
+docker compose build --no-cache
+docker compose up --exit-code-from unit-tests unit-tests
 
 test_result=$?